Refactor(WebSocket models)
