+++ src/services/contact/ContactService.ts
@@ -17,7 +17,7 @@ class ContactsService {
 
   /// #SERVER SERVICE CONNECTION
 
-  updateLocal = async (version: number): Promise<void> => {
+  updateLocal = async (newVersion: number): Promise<void> => {
     const request = await DinoAgentService.get(DinoAPIURLConstants.CONTACT_GET)
   
     if (request.canGo) {
@@ -28,7 +28,7 @@ class ContactsService {
         if (response.status === HttpStatus.OK) {
           const serverContacts: ContactModel[] = response.body
   
-          this.setVersion(version)
+          this.setVersion(newVersion)
   
           const newLocalItens = this.getItems().filter(c => c.id === undefined)
