--- src/types/app_settings/AppSettingsWebSocketAlertUpdateModel.ts
@@ -1,3 +1,0 @@
-export default interface AppSettingsWebSocketAlertUpdateModel {
-  newVersion: number
-}
