--- src/types/contact/ContactWebSocketAlertUpdateModel.ts
@@ -1,3 +1,0 @@
-export default interface ContactWebSocketAlertUpdateModel {
-    newVersion: number
-}
