--- src/types/faq/FaqUserWebSocketAlertUpdateModel.ts
@@ -1,3 +1,0 @@
-export default interface FaqUserWebSocketAlertUpdateModel {
-  newId: number
-}
