--- src/types/faq/FaqWebSocketAlertUpdateModel.ts
@@ -1,4 +1,0 @@
-export default interface FaqWebSocketAlertUpdateModel {
-    newId: number,
-    newVersion: number
-}
