--- src/types/glossary/GlossaryWebSocketAlertUpdateModel.ts
@@ -1,3 +1,0 @@
-export default interface GlossaryWebSocketAlertUpdateModel {
-  newVersion: number
-}
