--- src/types/note/NoteWebSocketAlertUpdateModel.ts
@@ -1,3 +1,0 @@
-export default interface NoteWebSocketAlertUpdateModel {
-  newVersion: number
-}
