--- src/types/user/UserAlertUpdateModel.ts
@@ -1,3 +1,0 @@
-export default interface UserAlertUpdateModel {
-  newVersion: number
-}
