+++ src/types/web_socket/WebSocketAlertUpdateModel.ts
@@ -1,0 +1,4 @@
+export default interface WebSocketAlertUpdateModel {
+    newId: number,
+    newVersion: number
+}
