+++ src/websocket/AppSettingsWebSocketSubscriber.ts
@@ -1,6 +1,6 @@
 import BaseWebSocketSubscriber from './BaseWebSocketSubscriber'
 import DinoAPIWebSocketConstants from '../constants/dino_api/DinoAPIWebSocketConstants'
-import AppSettingsWebSocketAlertUpdateModel from '../types/app_settings/AppSettingsWebSocketAlertUpdateModel'
+import WebSocketAlertUpdateModel from '../types/web_socket/WebSocketAlertUpdateModel'
 import AppSettingsService from '../services/app_settings/AppSettingsService'
 import SubscriberItem from '../types/web_socket/SubscriberItem'
 
@@ -8,7 +8,7 @@ class AppSettingsWebSocketSubscriber implements BaseWebSocketSubscriber {
   items: SubscriberItem[] = [
     {
       path: DinoAPIWebSocketConstants.ALERT_APP_SETTINGS_UPDATE,
-      callback: (model: AppSettingsWebSocketAlertUpdateModel) => {
+      callback: (model: WebSocketAlertUpdateModel) => {
         AppSettingsService.update(model.newVersion)
       },
     },
