+++ src/websocket/FaqWebSocketSubscriber.ts
@@ -1,6 +1,6 @@
 import BaseWebSocketSubscriber from './BaseWebSocketSubscriber'
 import DinoAPIWebSocketConstants from '../constants/dino_api/DinoAPIWebSocketConstants'
-import FaqWebSocketAlertUpdateModel from '../types/faq/FaqWebSocketAlertUpdateModel'
+import WebSocketAlertUpdateModel from '../types/web_socket/WebSocketAlertUpdateModel'
 import Service from '../services/faq/FaqService'
 import SubscriberItem from '../types/web_socket/SubscriberItem'
 
@@ -8,7 +8,7 @@ class FaqWebSocketSubscriber implements BaseWebSocketSubscriber {
   items: SubscriberItem[] = [
     {
       path: DinoAPIWebSocketConstants.ALERT_FAQ_UPDATE,
-      callback: (model: FaqWebSocketAlertUpdateModel) => {
+      callback: (model: WebSocketAlertUpdateModel) => {
         if(Service.getUserFaqId() === model.newId) { 
           if(Service.getVersion() !== model.newVersion) {
             Service.getUserFaqFromServer()
